Applied diff from #16003

+++ chia/seeder/crawl_store.py
@@ -11,6 +11,7 @@ from typing import Dict, List
 import aiosqlite
 
 from chia.seeder.peer_record import PeerRecord, PeerReliability
+from chia.util.ints import uint32, uint64
 
 log = logging.getLogger(__name__)
 
@@ -157,7 +158,7 @@ class CrawlStore:
         age_timestamp = int(max(peer.last_try_timestamp, peer.connected_timestamp))
         if age_timestamp == 0:
             age_timestamp = now - 1000
-        replaced = dataclasses.replace(peer, try_count=peer.try_count + 1, last_try_timestamp=now)
+        replaced = dataclasses.replace(peer, try_count=uint32(peer.try_count + 1), last_try_timestamp=uint64(now))
         reliability = await self.get_peer_reliability(peer.peer_id)
         if reliability is None:
             reliability = PeerReliability(peer.peer_id)
@@ -169,7 +170,7 @@ class CrawlStore:
         age_timestamp = int(max(peer.last_try_timestamp, peer.connected_timestamp))
         if age_timestamp == 0:
             age_timestamp = now - 1000
-        replaced = dataclasses.replace(peer, connected=True, connected_timestamp=now, tls_version=tls_version)
+        replaced = dataclasses.replace(peer, connected=True, connected_timestamp=uint64(now), tls_version=tls_version)
         reliability = await self.get_peer_reliability(peer.peer_id)
         if reliability is None:
             reliability = PeerReliability(peer.peer_id)
